d/control: Add ${gir:Provides} to -dev packages
With a new enough gobject-introspection, this adds Provides on systematic
names gir1.2-x-y-dev that make it easier to handle GObject-Introspection
dependencies uniformly, and will eventually make it possible to exclude
GIR XML during cross-compilation.

In this particular package, the Provides is mainly so that GTK 2 and 3
can add Build-Depends on gir1.2-atk-1.0-dev, indicating that their
GObject-Introspection build requires GIR XML from at-spi2-core.

With older gobject-introspection, this change has no effect (but is
harmless).

Helps: #1030223