Build-depend on dbus-daemon and mark as nocheck
The dbus package pulls in everything needed for a system bus, but the
package is only needed for tests, and to find the path to the binary.
Tests work with the simpler dbus-daemon package, and the path to the
binary is fixed so just add it via the apposite meson option to avoid
dynamic detection and the need for a build dependency.